change build system to cmake
The build system has been changed to CMake in order to support builds on both
Windows and Linux systems with no changes to the build system. This change also
introduces the config portion of the source files which can be used to separate
sources based on the availability of headers, libraries, or anything else.