add facility and severity support (#8)
Facility and severity values can be specified in two places. The first is as a default value for a given target, which is used when the target is called with the shortened version of call without a specific entry. The second is for an entry which has the values provided.

These values are also made to be compatible with `syslog.h`, if it exists. This means that the underlying values of `#define`s such as `STUMPLESS_FACILITY_USER` will be that of the `syslog.h` value (`LOG_USER` for the previous example). If the standard `syslog.h` header is not available, values are used that are compliant with RFC 5424.