add default app name and msgid to target (#10)
Previously the app name and msgid portions of log messages created using the
`stumpless` function call were hard-coded values. This change introduces
default values for these fields that are used for the simplified `stumpless`
call when an entry is not used. If these defaults are not used, the RFC
specified `-` character is used instead.