bring coverage up to 80% (#12)
Add new unit tests in order to bring the code coverage up to 80% per the
project's new standards. These tests have identified some errors in the code
base which are also fixed in this change.

This change also begins the use of the SPDX identifier in sourc efiles in order
to allow fo external tools to identify the license in use.