fix broken google test builds (#25)
Google Test builds broke due to pulling the master branch of the repository,
which changed the location of the libraries as well as required manual use of
the C++11 standard. The version of Google Test is now specified using the
release tag of the repository, currently 1.8.1. Moving forward, versions of
external libraries need to be manually specified and updated to keep project
builds stable.