add format specifier support (#43)
In order to be a viable replacement for the syslog function, the library will
need to support format specifiers in messages. This is added to not only the
simple logging calls, but also entry creation as well. In addition, variadic
arguments allows insertion strings in Windows Event Log targets to be set in
a less verbose manner that aligns closely with the format specifier approach.