make network examples run manually (#48)
Network examples (UDP and TCP) are currently responsible for a large portion
of time spent in the CI builds, as well for as generating unnecessary network
traffic that is not actually needed. By removing these examples from the
examples build target so that they are only run manually, presumably after
they've been modified to do something meaningful, the CI speed can be
significantly increased.