add default target capability (#49)
The standard library syslog function will send messages to a default location
if a log has not yet been opened. This same behavior is added here, as well as
the stumplog function, which has an identical signature to the syslog function
and is intended to function as an easy way to transition into using stumpless.