support mingw and cygwin (#61)
Builds on mingw and cygwin fail at several points due to issues in the
build scripts and mistaken assumptions about the environment. This is
corrected by increasing the portability of the code in points where
environment issues cause the problems, and updating the build scripts where
they are the culprit.

A number of portability improvements are also included here, including
updates to the most recent (and working) Google Test and Benchmark versions
and use of the sysconf function when it is available.