add c++ library (#66)
Add C++ language bindings. These are implemented using Wrapture and a set
of class specifications.

A Gemfile has been added to the project to support easy installation of the
new dependency for developers. Future tools will be preferred if they are
Ruby gems that can be installed this same way.