modify linguist detection (#71)
Linguist had labelled the repository's main language as C++ due to the
amount of test code being greater than the actual library code. In order to
prevent this, `.cpp` and `.hpp` files have been marked as documentation
within linguist to suppress their effect on the repository language
statistics.

While this is not an ideal solution, it does cause Github to classify the
repository as a C project. This will be revisited if an alternative
solution presents itself.