rename default branch (#72)
The branch scheme has been changed to have two long-lived branches. The
default branch is now named `latest` and serves as the base for any new
feature or change branch. The other branch, `release`, is kept at the most
recent semantic version release and serves as a more stable alternative to
`latest` for anyone building from the source repository.

The Code of Conduct has also been updated to the latest version of the
Contributor Covenant, 2.0.

The renaming of the `master` branch to `latest` also serves to remove
language that may not be inclusive of all individuals. Stumpless seeks to
make participation in its community inclusive and harassment-free, as
detailed in the Code of Conduct. The project maintainers will work to make
any changes necessary, such as this one, in order to keep this pledge.