update local socket name generation (#76)
Socket targets or opened with a local socket name as an argument, which may
be set to `NULL`, in which case a name is generated (see the documentation
of `stumpless_open_socket_target` for details). This name is now generated
using a call to `mkstemp` to avoid naming conflicts with improperly
terminated socket targets.

This fixes #54, which can be referenced for more details.