fix closure of current target (#85)
The current target is set to either the last opened target, or with
`stumpless_set_current_target`. However when the target that is currently
designated as current is closed the current target pointer was not updated,
and as a result pointed to invalid memory. This change corrects this
behavior and adds tests for regression.

This fixes #52, which can be referenced for more details.