change version to 2.0.0 (#96)
Update the version of the code to 2.0.0 as features and changes from the
2.0.0 roadmap are added.

Note that until version 2.0.0 is released, changes scheduled for 2.0.0 will
be immediately visible in the `latest` branch as soon as they are
implemented, which may cause breakage for some users. If more stability is
needed, please use a specific release tag (e.g. `v1.6.0) or the `release`
branch until 2.0.0 is finished to avoid introducing breaking changes into
your application.