improve stumpless_copy_element performance (#97)
The `stumpless_copy_element` function made several calls to `realloc` in
a single invocation, once for each param in the original element. This was
inefficient, and has been changed to allocate the array for params once
before each param is copied, reducing the number of memory management calls
and the number of error checks needed during execution.

This change also adds detailed documentation on performance testing in
Stumpless. This doc is intended to serve as a guide for anyone wishing to
make a performance enhancement change.