add error id to stumpless_perror output (#99)
The error id was not included in the output of `stumpless_perror`, which
added extra steps when troubleshooting if a user needed to track it down.
This change adds it to these statements, and also provides a new function
to provide the string of any error id, `stumpless_get_error_id_string`.

This change also introduces a macro-based technique for keeping enum
definitions and their string descriptions in sync. This technique will
likely be adopted for existing and future enums in the library in order
to provide the same string representation if desired.

This fixes #53, which can be referenced for more details.