cleanup sockets leftover after tests (#104)
When running the test suite, several sockets were left in the file system
due to a failure of certain tests to clean up after opening the default
target as a byproduct of some assertions. This change adds the necessary
calls to clean these up. It also updates the documentation of
`stumpless_free_all` to make it clear that this function closes the default
target if it has been opened.