remove stumpless function (#106)
The `stumpless` function made it impossible for the C++ bindings to use
the most direct namespace name of `stumpless`. For this reason the
function has been removed along with the matching `vstumpless` function.
These functions should be replaced with the `stump` and `vstump` functions
introduced in v1.6.0.

The C++ namespace has also been changed from `stumplesscpp` to `stumpless`,
and the documentation has also been updated to reflect the new namespace.