clean up includes (#109)
This change makes a number of improvements to the include strategy.

Facility, option, and severity values were previously available in the
`stumpless/entry.h` header as a result of this being the file they were
original defined in. They were late split into their own files
(`stumpless/facility.h`, stumpless/option.h`, and `stumpless/severity.h`),
but these were left in the original header as well so that older code
would continue to compile. Version 2.0.0 moves them out of the entry file
completely.

Stumpless also previously defined `WIN32_LEAN_AND_MEAN` whenever it included
`windows.h` in the `stumpless/config.h` header. While this did not cause any
reported problems, it did have the potential to cause issues or difficult to
understand behavior if used in a project which needed this symbol to not be
defined. The definition has now been removed, and use of Windows headers has
been moved to a wrapper file to address inclusion ordering issues that were
uncovered by the change.