minor documentation tweaks (#112)
Several contributors have followed the guidance in the
`docs/CONTRIBUTING.md` document which states that PRs should keep the
number of commits to a minimum. However doing this requires force-pushing
each change to a PR so that it remains a single commit, which makes it
difficult to track changes to the PR as it develops. This change softens
the language of the contribution guidelines to make it clear that smaller
changes (such as first issue ones) do not need to worry about commit
history cleanliness as they will be squash merged.

The current structure of the option header `include/stumpless/option.h`
required the documentation of the option to be repeated in two places in
the same file during implementation. Since this is a clear violation of the
DRY principle and is no fault of the implementer of the option, the header
has been restructured to only document each option once. This uses the same
strategy as the include/stumpless/level headers which had similar issues
during their development. This will simplify the documentation step of the
remaining open issues for option implementation.

This change also corrects a broken email link in the main README.