add localization framework (#115)
Stumpless has a number of strings that are presented to users, such as
error messages. In order to make the library more accessible, these
have been localized so that they can be presented in an appropriate
language for the context.

This change adds a localization framework for these strings as well as
documentation for working with it. Existing strings have been moved to the
en-US locale, and a stub fr-FR locale has also be included to ensure that
different locales are handled properly. CI tests for both Linux and Windows
have also been added to test locales different from the default.

In support of this framework a validation module has also been added in
`include/private/validate.h` which includes a macro function for testing if
arguments are NULL.