implement stumpless_param_to_string (#134)
It is often useful to be able to obtain out a string representation of a
given object for user interface hints, troubleshooting, or operations
such as value comparison. This change implements the
`stumpless_param_to_string` function to for param structs.

This fixes #128, which can be referenced for more details.