add benchmarks to ci (#135)
The benchmark for the `stumpless` function had not been changed to the new
`stump` function, and as a result was failing. This change updates the
benchmark with the correct function name, adds the `bench` target to the
Travis CI Release builds so that future problems are caught during CI, and
fixes other issues exposed by the new CI builds.