implement thread safety (#136)
This change implements thread safety for all possible library operations.
Some functions where this is not feasible still need to be synchronized
to some extent, such as destructor functions and other memory management
routines.

This is a large change in both size and complexity, and can cause
breakage in some cases. The buffer target interface has changed so that it
can accommodate multithreaded reads, and functions returning string values
now require their result to be freed when it is no longer needed. More
details on these items can be found in the updates to the `Changelog.md`
file in this commit.