remove rarely used parameters (#144)
Functions to open targets always required parameters for options and the
default facility, as they were based on the `syslog.h` `openlog` function
signature. However, both parameters are rarely actually used for anything
other than boilerplate settings, and have therefore been removed to allow
code to be more concise. If either of these fields need to be adjusted on
a target, use the `stumpless_set_option` and
`stumpless_set_default_facility` functions on the target after it has been
created.