deprecate ambiguous destructors (#148)
There are two forms of the destructors for entry and element structures:
one that destroys the object itself, and one that destroys the object and
all of the ones that it contains. The former is named `destroy_..._only`
while the latter is named  `destroy_..._and_contents`. The latter has a
synonym named simply `destroy_...`, which does not convey the nuance of
its behavior at all. This alias has therefore been deprecated, and will
be removed in release 3.0.0 in order to prevent confusion and misuse of
the two forms.