add missing c++ bindings (#150)
There are some functions that did not have corresponding bindings in the
C++ wrappers generated by wrapture. This change adds specs for these
functions, and also updates wrapture to allow error checking and return
codes to be handled properly in them.

For an exact list of the new functions, see the updates to `ChangeLog.md`
in this update.