fix NULL dereferences in stumpless_perror (#178)
Several errors did not have messages, which led to NULL dereferences in
`stumpless_perror` when an attempt was made to print them. Localized
error messages have been added to each of these and tests added (where
possible) to catch regression. Refer to the Changelog for a full list of
affected errors.