move mutexes to structure definitions (#204)
In the initial implementation of multithreading support, the mutex used to
protect target, entry, element, and param types was stored at the end of
the struct in an effort to avoid a dereference operation and simplify
initialization.

Unfortunately, this leads to a mismatch in the memory required for a proper
target structure and the size of the structure according to the compiler.
This in turn can lead to undefined behavior when using a stack-allocated
structure. While this isn't currently supported usage, it is planned for
future releases, and is also used by some test suites (unsupported target
types being one) which leads to instability.

This change adds a void pointer to the end of the structure instead which
points at the mutex to use for synchronization of these types. These are
allocated using a dedicated slab cache instead of as part of the struct's
initial allocation.

This fixes #203, which can be referenced for more details.