add tracing capabilities (#209)
Add several tracing functions that log entries along with the file, line
number, and function of the source code. These can be used by either
specifying the trace information in the function call, or using a macro
that automatically includes it. The messages themselves have a severity
of debug.

This also includes a number of improvements of the target test suite in
order to increase coverage, along with corrected memory leaks discovered
during this testing. The most notable of these is a leak that occurred
within a specific memory allocation failure scenario in the slab caching
mechanism.