fix documentation and journald tests (#220)
Add some clarifications to the documentation that have been missed in
recent contributions, such as the since tag to denote when functions
were added and alphabetization of functions.

Also add recently contributed localizations to the changelog for the
upcoming version.

Journald function tests are also failing periodically, likely due to
the check not running for long enough for the log to come through to
the journald search. The search attempt count has been doubled in an
attempt to fix this.