correct perror multithreading issue (#235)
Protects writes to the error stream that occur as a result of the
STUMPLESS_OPTION_PERROR target option. This is done using the error
stream lock, so that they are no longer unsafe in multithreaded
programs. A number of tests are also included to prevent regressions
of this in the future.

This resolves #207, which can be referenced for more details.

Co-authored-by: JatishKharade <56515325+Jk-bit@users.noreply.github.com>
Co-authored-by: igorkole <90953863+igorkole@users.noreply.github.com>