add string to severity and facility functions (#239)
While functions exist to get a string representation of either a
severity or facility enumeration, it was not possible to go in opposite
direction. This change adds functions that will attempt to convert a
provided string into a severity or facility enumeration value.

This resolves #237, which can be referenced for more details.