add target type to string function (#262)
The target type enumeration was not defined using the foreach macro
function pattern of other enumerations, meaning that a to string function
would have been written manually. This change redefines the enumeration
using the macro function style, and adds a to string function using this.

This fixes #246, which can be referenced for more details.