avoid SIGPIPE termination (#260)
Adds a MSG_NOSIGNAL flag to POSIX network send calls to avoid program
termination via a SIGPIPE signal. While this was not an issue on
Windows builds, Linux systems at a minimum were affected.

A roadmap item has been added for the next minor release to improve error
handling so that a closed session is detected faster than a failed call
to send, which may take longer than necessary in some cases.

This fixes #231, which can be referenced for more details.