allow procid override (#267)
Previously the procid field of entries was derived from the process ID.
However, the field is a 128 printable ASCII field in RFC 5424, and users
may want to set this id to something more meaningful to them. This change
allows this default behavior to be overriden.

This fixes #251, which can be referenced for more details.