refactor and improve coverage (#313)
Fixes remaining sonarcloud violations not associated with todos.

Adds tests to slightly improve coverage, as well as updates the github
action used to upload coverage. Reported coverage on the project has
decreased significantly as a result of this update, seemingly because of
a difference in the way conditional statements are counted.

Removes some now-obsolete #define values.