fix mingw build and docs (#320)
Fixes warnings encountered during MinGW builds due to type mismatches.
Also updates the build documentation to include more specific instructions
for MinGW platforms.

This partially resolves #318, which can be referenced for more details.
It does not fix the build on older versions of MinGW, which will be
handled in a separate change if needed.

The additional information on building on MinGW and MSYS2 may also
help #303.