add load constructors (#321)
Often a structure can be stored on the stack instead of the heap, which
can in turn lead to better performance. This change adds support for
stack allocation of entry, element, and param structures. It also makes
use of this capability for some internally stored structures, in an effort
to reduce heap memory usage and improve performance.