add ansi c support (#328)
Library compilation and functionality was not portable enough to support
builds in ANSI C only environments, where some functions such as gmtime_r
and getaddrinfo are not provided. This change updates the configuration
and build to support these environments, and improves documentation in the
affected areas of functionality to more clearly convey caveats and
limitations.

This fixes #325, which can be referenced for more details.