add fuzzing test framework (#329)
Fuzzing tests provide an avenue to identify potential security issues in
functionality. This change adds a framework to write and run fuzzers against
stumpless, including a fuzzer for the stumpless_add_message_str function.

Corpus directories are also included, and their contents can be used in unit
tests as well as fuzzing. This is intended to facilitate both testing
against problems that a fuzzer finds and improving the seed for future
fuzzing.