add ebuild support (#355)
Adds an ebuild template that is generated at configuration time and can be
used to install the library on Gentoo systems.

New build targets and configuration options are also included in this change
to improve installation support, not just for Gentoo but in general. These
include options to install manpages and examples and a build-test target to
compile functionality tests without running them.