add sqlite3 targets (#393)
Adds SQLite3 targets for logging to a SQLite database.

The indent wrapper script has been removed as part of this change, as
some of the formatting that it enforces is undesirable. Specifically,
the indention of #define statements within header guards, and function
calls with no parameters having two spaces inside will no longer be used
in new code moving forward.