use library strncasecmp if available
Introduces a new configuration check for the availability of strncasecmp
in strings.h, only using a custom implementation if it is not available.

This change uncovered a bug in stumpless_get_severity_enum_from_buffer,
which does not return an error for incomplete severity names as is. This
will be fixed in a future change.