Improve journald initialization
The current implementation of the journald target creates a copy of the
`fixed_fields` structure for each message. This is inefficient and
results in a temporary memory leak that valgrind picks up.

This change switches to a thread local (where available) static
struct which elimiates the need to copy the memory or free it upon
completion.

This change also remove the valgrind suppressions for this problem.