refactor journald initialization
The current implementation of the journald target creates a dynamic
fixed_fields structure used for each message. This is inefficient and
contains a memory use issue that valgrind picks up.

This change switches to a thread local (where available) static struct
which elimiates the need to copy the memory or free it upon completion.
It also removes the valgrind suppressions for this problem.